Js refactor
