+++ webapp/src/main/resources/static/js/admin-page/adminPage.js
@@ -1,4 +1,4 @@
-import {getUsers} from "./ajax/userRestController.js";
+import {getUsers} from "../ajax/userRestController.js";
 
 window.addEventListener('load', function () {
