--- webapp/src/main/resources/static/js/menu-page/components/menu-box/navbar/navbar.js
@@ -1,3 +1,0 @@
-/**
- * Created by Михаил on 18.11.2019.
- */
