--- webapp/src/main/resources/static/js/menu-page/components/navbar/page-navbar.js
@@ -1,3 +1,0 @@
-/**
- * Created by Михаил on 18.11.2019.
- */
