+++ webapp/src/main/resources/static/js/sign-in-page/sign-in-page.js
@@ -1,4 +1,4 @@
-import {WorkspaceRestPaginationService} from './rest/entities-rest-pagination.js'
+import {WorkspaceRestPaginationService} from '../rest/entities-rest-pagination.js'
 
 
 const workspace_service = new WorkspaceRestPaginationService();
