+++ webapp/src/main/resources/static/js/workspace-page/components/footer/input-box-at.js
@@ -1,5 +1,5 @@
 export let users = [];
-import {UserRestPaginationService} from './rest/entities-rest-pagination.js'
+import {UserRestPaginationService} from '../../../rest/entities-rest-pagination.js'
 
 const user_service = new UserRestPaginationService();
