+++ webapp/src/main/resources/static/js/workspace-page/workspace-page.js
@@ -1,5 +1,5 @@
-import {ChannelRestPaginationService, BotRestPaginationService, WorkspaceRestPaginationService} from './rest/entities-rest-pagination.js'
-import {getAllUsersInThisChannel} from "./ajax/userRestController.js";
+import {ChannelRestPaginationService, BotRestPaginationService, WorkspaceRestPaginationService} from '../rest/entities-rest-pagination.js'
+import {getAllUsersInThisChannel} from "../ajax/userRestController.js";
 
 const channel_service = new ChannelRestPaginationService();
 const bot_service = new BotRestPaginationService();
