+++ webapp/src/main/resources/templates/admin-page.html
@@ -5,9 +5,9 @@
     <title>Admin Page</title>
     <meta charset="utf-8">
 
-    <link rel="stylesheet" th:href="@{/css/adminPage.css}"/>
+    <link rel="stylesheet" th:href="@{/js/admin-page/adminPage.css}"/>
 
-    <script type="module" th:src="@{/js/adminPage.js}"></script>
+    <script type="module" th:src="@{/js/admin-page/adminPage.js}"></script>
 
     <!-- Bootstrap CSS -->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css"
