+++ webapp/src/main/resources/templates/choose-workspace-page.html
@@ -3,9 +3,9 @@
 <head>
     <meta charset="UTF-8">
     <title>choose workspace</title>
-    <link rel="stylesheet" type="text/css" href="css/chooseWorkspacePage.css">
-    <script src="/js/jquery-3.4.1.min.js"></script>
-    <script type="module" src="/js/choose-workspace.js"></script>
+    <link rel="stylesheet" type="text/css" href="js/choose-workspace-page/chooseWorkspacePage.css">
+    <script src="/js/jquery/jquery-3.4.1.min.js"></script>
+    <script type="module" src="/js/choose-workspace-page/choose-workspace.js"></script>
 </head>
 <body>
 <label for="favcity">
