+++ webapp/src/main/resources/templates/createWorkspace/channel-name.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <title>Create a Workspace</title>
     <link rel="stylesheet" type="text/css" href="css/CreateWorkspaceChannelNamePage.css">
-    <script src="/js/jquery-3.4.1.min.js"></script>
+    <script src="/js/jquery/jquery-3.4.1.min.js"></script>
     <script type="module" src="/js/createWorkspace/channel-name.js"></script>
 </head>
 <body>
