+++ webapp/src/main/resources/templates/createWorkspace/new-workspace-page.html
@@ -3,7 +3,7 @@
 <head>
     <title>Create a workspace</title>
     <link rel="stylesheet" type="text/css" href="css/newWorkspacePage.css">
-    <script src="/js/jquery-3.4.1.min.js"></script>
+    <script src="/js/jquery/jquery-3.4.1.min.js"></script>
     <script type="module" src="js/createWorkspace/new-wrokspace.js"></script>
     <meta charset="utf-8">
 </head>
