+++ webapp/src/main/resources/templates/search-channel-page.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Workspace Page</title>
-    <script th:src="@{/js/jquery-3.4.1.min.js}"></script>
+    <script src="/js/jquery/jquery-3.4.1.min.js"></script>
     <script type="module" th:src="@{/js/search-channel-page/searchChannel.js}"></script>
     <link rel="stylesheet" th:href="@{/js/search-channel-page/searchChannel.css}"/>
 </head>
