+++ webapp/src/main/resources/templates/search-users-page.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Workspace Page</title>
-    <script th:src="@{/js/jquery-3.4.1.min.js}"></script>
+    <script src="/js/jquery/jquery-3.4.1.min.js"></script>
     <script type="module" th:src="@{/js/search-users-page/searchUsers.js}"></script>
     <link rel="stylesheet" th:href="@{/js/search-users-page/searchUsers.css}"/>
 </head>
