+++ webapp/src/main/resources/templates/signin-page.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <title>Sign In Page</title>
     <link rel="stylesheet" th:href="@{/js/sign-in-page/signInPage.css}"/>
-    <script src="/js/jquery/-3.4.1.min.js"></script>
+    <script src="/js/jquery/jquery-3.4.1.min.js"></script>
     <script type="module" src="/js/sign-in-page/sign-in-page.js"></script>
 </head>
 <body>
