+++ webapp/src/main/resources/templates/workspaces.html
@@ -3,12 +3,12 @@
 <head>
     <meta charset="UTF-8">
     <title>Title</title>
-    <script src="/js/jquery-3.4.1.min.js"></script>
+    <script src="/js/jquery/jquery-3.4.1.min.js"></script>
     <script type="module" src="/js/workspaces-page/workspaces.js"></script>
     <script type="module">update_workspace_table()</script>
     <script type="module">update_add_new_workspace_table()</script>
     <script src="/js/helper_functions.js"></script>
-    <script src="../static/js/workspaces-page/workspaces.js"></script>
+    <!--<script src="../static/js/workspaces-page/workspaces.js"></script>-->
 </head>
 <body>
 <table border="5" id="table_update_delete_worspaces">
